src/requirements.txt
numpy
onnxruntime
pillow
pyyaml